Library upgrades: springframework , hibernate, quartz. Security library upgrade required code changes added 2 features too while there. Audit on session expiration and max session of 1 per user.

+++ core/src/main/java/org/akaza/openclinica/dao/hibernate/AuditUserLoginDao.java
@@ -24,7 +24,7 @@ public class AuditUserLoginDao extends AbstractDomainDao<AuditUserLoginBean> {
         Criteria criteria = getCurrentSession().createCriteria(domainClass());
         criteria = filter.execute(criteria);
         criteria.setProjection(Projections.rowCount()).uniqueResult();
-        return (Integer) criteria.uniqueResult();
+        return ((Long) criteria.uniqueResult()).intValue();
     }
 
     @SuppressWarnings("unchecked")
